# api-spec.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: |
    API para gestión de tareas siguiendo el patrón API First.
    
    Esta API permite crear, leer, actualizar y eliminar tareas,
    así como gestionar usuarios y categorías.
  version: 1.0.0
  contact:
    name: API Support
    email: support@taskapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000/api/v1
    description: Servidor de desarrollo
  - url: https://api.taskmanagement.com/v1
    description: Servidor de producción

paths:
  /health:
    get:
      summary: Health check del servicio
      description: Verifica que el servicio esté funcionando correctamente
      operationId: healthCheck
      tags:
        - Health
      responses:
        '200':
          description: Servicio funcionando correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /tasks:
    get:
      summary: Obtener todas las tareas
      description: Recupera una lista paginada de todas las tareas
      operationId: getTasks
      tags:
        - Tasks
      parameters:
        - name: page
          in: query
          description: Número de página
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Número de elementos por página
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          description: Filtrar por estado
          required: false
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: category_id
          in: query
          description: Filtrar por categoría
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Lista de tareas recuperada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksResponse'
        '400':
          description: Parámetros de consulta inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Crear una nueva tarea
      description: Crea una nueva tarea en el sistema
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Tarea creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  /tasks/{taskId}:
    get:
      summary: Obtener una tarea específica
      description: Recupera los detalles de una tarea por su ID
      operationId: getTaskById
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: ID único de la tarea
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Tarea encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Actualizar una tarea
      description: Actualiza completamente una tarea existente
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: ID único de la tarea
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Tarea actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

    patch:
      summary: Actualizar parcialmente una tarea
      description: Actualiza parcialmente una tarea existente
      operationId: patchTask
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: ID único de la tarea
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTaskRequest'
      responses:
        '200':
          description: Tarea actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Eliminar una tarea
      description: Elimina permanentemente una tarea del sistema
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: ID único de la tarea
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Tarea eliminada exitosamente
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /categories:
    get:
      summary: Obtener todas las categorías
      description: Recupera una lista de todas las categorías disponibles
      operationId: getCategories
      tags:
        - Categories
      responses:
        '200':
          description: Lista de categorías recuperada exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      summary: Crear una nueva categoría
      description: Crea una nueva categoría en el sistema
      operationId: createCategory
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Categoría creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  /categories/{categoryId}:
    get:
      summary: Obtener una categoría específica
      description: Recupera los detalles de una categoría por su ID
      operationId: getCategoryById
      tags:
        - Categories
      parameters:
        - name: categoryId
          in: path
          required: true
          description: ID único de la categoría
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Categoría encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Actualizar una categoría
      description: Actualiza una categoría existente
      operationId: updateCategory
      tags:
        - Categories
      parameters:
        - name: categoryId
          in: path
          required: true
          description: ID único de la categoría
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Categoría actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Eliminar una categoría
      description: Elimina permanentemente una categoría del sistema
      operationId: deleteCategory
      tags:
        - Categories
      parameters:
        - name: categoryId
          in: path
          required: true
          description: ID único de la categoría
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Categoría eliminada exitosamente
        '404':
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: No se puede eliminar la categoría porque tiene tareas asociadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - description
        - status
        - priority
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: ID único de la tarea
          example: 1
        title:
          type: string
          description: Título de la tarea
          maxLength: 200
          example: "Completar documentación de API"
        description:
          type: string
          description: Descripción detallada de la tarea
          maxLength: 1000
          example: "Escribir la documentación completa de la API REST incluyendo ejemplos"
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        due_date:
          type: string
          format: date-time
          description: Fecha límite para completar la tarea
          example: "2024-12-31T23:59:59Z"
        category_id:
          type: integer
          description: ID de la categoría asociada
          example: 1
        category:
          $ref: '#/components/schemas/Category'
        created_at:
          type: string
          format: date-time
          description: Fecha y hora de creación
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Fecha y hora de última actualización
          example: "2024-01-15T14:45:00Z"

    TaskStatus:
      type: string
      enum:
        - pending
        - in_progress
        - completed
        - cancelled
      description: Estado actual de la tarea
      example: "pending"

    TaskPriority:
      type: string
      enum:
        - low
        - medium
        - high
        - urgent
      description: Prioridad de la tarea
      example: "medium"

    Category:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: ID único de la categoría
          example: 1
        name:
          type: string
          description: Nombre de la categoría
          maxLength: 100
          example: "Desarrollo"
        description:
          type: string
          description: Descripción de la categoría
          maxLength: 500
          example: "Tareas relacionadas con desarrollo de software"
        color:
          type: string
          description: Color hexadecimal para la categoría
          pattern: "^#[0-9A-Fa-f]{6}$"
          example: "#3498db"
        created_at:
          type: string
          format: date-time
          description: Fecha y hora de creación
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Fecha y hora de última actualización
          example: "2024-01-15T14:45:00Z"

    CreateTaskRequest:
      type: object
      required:
        - title
        - description
        - priority
      properties:
        title:
          type: string
          description: Título de la tarea
          maxLength: 200
          example: "Completar documentación de API"
        description:
          type: string
          description: Descripción detallada de la tarea
          maxLength: 1000
          example: "Escribir la documentación completa de la API REST incluyendo ejemplos"
        priority:
          $ref: '#/components/schemas/TaskPriority'
        due_date:
          type: string
          format: date-time
          description: Fecha límite para completar la tarea
          example: "2024-12-31T23:59:59Z"
        category_id:
          type: integer
          description: ID de la categoría asociada
          example: 1

    UpdateTaskRequest:
      type: object
      required:
        - title
        - description
        - status
        - priority
      properties:
        title:
          type: string
          description: Título de la tarea
          maxLength: 200
          example: "Completar documentación de API"
        description:
          type: string
          description: Descripción detallada de la tarea
          maxLength: 1000
          example: "Escribir la documentación completa de la API REST incluyendo ejemplos"
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        due_date:
          type: string
          format: date-time
          description: Fecha límite para completar la tarea
          example: "2024-12-31T23:59:59Z"
        category_id:
          type: integer
          description: ID de la categoría asociada
          example: 1

    PatchTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Título de la tarea
          maxLength: 200
          example: "Completar documentación de API"
        description:
          type: string
          description: Descripción detallada de la tarea
          maxLength: 1000
          example: "Escribir la documentación completa de la API REST incluyendo ejemplos"
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        due_date:
          type: string
          format: date-time
          description: Fecha límite para completar la tarea
          example: "2024-12-31T23:59:59Z"
        category_id:
          type: integer
          description: ID de la categoría asociada
          example: 1

    CreateCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Nombre de la categoría
          maxLength: 100
          example: "Desarrollo"
        description:
          type: string
          description: Descripción de la categoría
          maxLength: 500
          example: "Tareas relacionadas con desarrollo de software"
        color:
          type: string
          description: Color hexadecimal para la categoría
          pattern: "^#[0-9A-Fa-f]{6}$"
          example: "#3498db"

    UpdateCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Nombre de la categoría
          maxLength: 100
          example: "Desarrollo"
        description:
          type: string
          description: Descripción de la categoría
          maxLength: 500
          example: "Tareas relacionadas con desarrollo de software"
        color:
          type: string
          description: Color hexadecimal para la categoría
          pattern: "^#[0-9A-Fa-f]{6}$"
          example: "#3498db"

    TasksResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    PaginationInfo:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          description: Página actual
          example: 1
        limit:
          type: integer
          description: Elementos por página
          example: 10
        total:
          type: integer
          description: Total de elementos
          example: 25
        total_pages:
          type: integer
          description: Total de páginas
          example: 3
        has_next:
          type: boolean
          description: Indica si hay una página siguiente
          example: true
        has_prev:
          type: boolean
          description: Indica si hay una página anterior
          example: false

    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - version
      properties:
        status:
          type: string
          description: Estado del servicio
          example: "ok"
        timestamp:
          type: string
          format: date-time
          description: Timestamp del health check
          example: "2024-01-15T10:30:00Z"
        version:
          type: string
          description: Versión de la API
          example: "1.0.0"
        uptime:
          type: number
          description: Tiempo de actividad en segundos
          example: 3600

    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Código de error
          example: "NOT_FOUND"
        message:
          type: string
          description: Mensaje descriptivo del error
          example: "La tarea solicitada no fue encontrada"
        timestamp:
          type: string
          format: date-time
          description: Timestamp del error
          example: "2024-01-15T10:30:00Z"
        details:
          type: object
          description: Detalles adicionales del error
          additionalProperties: true

    ValidationErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - validation_errors
      properties:
        error:
          type: string
          description: Código de error
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Mensaje descriptivo del error
          example: "Los datos proporcionados no son válidos"
        timestamp:
          type: string
          format: date-time
          description: Timestamp del error
          example: "2024-01-15T10:30:00Z"
        validation_errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Campo que contiene el error
                example: "title"
              message:
                type: string
                description: Mensaje de error para el campo
                example: "El título es requerido"
              code:
                type: string
                description: Código específico del error de validación
                example: "REQUIRED"

tags:
  - name: Health
    description: Endpoints para verificar el estado del servicio
  - name: Tasks
    description: Operaciones CRUD para gestión de tareas
  - name: Categories
    description: Operaciones CRUD para gestión de categorías